CollectionViewのFlowLayoutをそれぞれファイル分け
